requirements.txt
appdirs==1.4.4
asgiref==3.4.1
black==21.7b0
click==8.0.1
Django==3.2.6
django-ckeditor==6.1.0
django-cors-headers==3.7.0
django-environ==0.4.5
django-js-asset==1.2.2
django-taggit==1.5.1
django-taggit-serializer==0.1.7
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
drf-nested-routers==0.93.3
huey==2.4.0
mypy-extensions==0.4.3
pathspec==0.9.0
psycopg2-binary==2.9.1
PyJWT==2.1.0
python-http-client==3.3.2
python-slugify==5.0.2
pytz==2021.1
redis==3.5.3
regex==2021.8.3
sendgrid==6.8.0
six==1.16.0
sqlparse==0.4.1
starkbank-ecdsa==1.1.1
text-unidecode==1.3
tomli==1.2.1
